app/requirements.txt
google-api-python-client==2.36.0
psycopg2-binary==2.9.1
schedule==1.1.0